refactor(auth): 重命名ApiKey认证为AccessToken认证
1.  替换请求头中的X-Api-Key为X-Access-Token
2.  删除旧的ApiKey认证相关代码，新增Token认证处理器
3.  简化配置文件，移除旧的WebUI和日志相关配置，仅保留AccessToken配置
4.  简化服务器启动时的URL配置逻辑
