+++ SRAFrontend/SRAFrontend.Server/appsettings.json
@@ -6,17 +6,5 @@
     }
   },
   "AllowedHosts": "*",
-  "ApiKey": "starrailassistant",
-  "WebUi": {
-    "Enabled": true,
-    "Host": "0.0.0.0",
-    "Port": 5074,
-    "RemoteAccess": true,
-    "AuthToken": "starrailassistant",
-    "ReadonlyMode": false
-  },
-  "Logs": {
-    "TailLines": 300,
-    "StreamEnabled": true
-  }
+  "AccessToken": "starrailassistant"
 }
