+++ SRAFrontend/srafrontend-webui/src/styles/app.css
@@ -147,7 +147,7 @@ select {
     background 180ms ease,
     border-color 180ms ease,
     box-shadow 180ms ease,
-    transform 200ms ease;
+    transform 180ms ease;
 }
 
 .panel:hover,
